data_pipeline/requirements.txt
alpaca-trade-api==3.0.0
python-binance==1.0.19
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
yfinance==0.2.18
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.65.0
